+++ src/framework/piral-base/src/cleanup.ts
@@ -1,3 +1,4 @@
+import { getBasePath } from './utils';
 import { SinglePilet } from './types';
 
 export function cleanup(pilet: SinglePilet) {
@@ -8,5 +9,15 @@ export function cleanup(pilet: SinglePilet) {
   if ('requireRef' in pilet) {
     const depName = pilet.requireRef;
     delete window[depName];
+  }
+
+  // remove the pilet's evaluated modules from SystemJS (except the shared dependencies)
+  if ('link' in pilet) {
+    const basePath = getBasePath(pilet.link);
+    const dependencies = Object.keys(pilet.dependencies || {}).map((m) => pilet.dependencies[m]);
+
+    [...System.entries()]
+      .filter(([id]) => id.startsWith(basePath) && !dependencies.includes(id))
+      .forEach(([id]) => System.delete(id));
   }
 }