+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Updated dependencies
 - Fixed issue with `pilet build` of `--type standalone` when using importmap
 - Added `CUSTOM_ELEMENTS_SCHEMA` to default / fallback module in `piral-ng`
+- Added automatic module cleanup on pilet teardown
 
 ## 0.14.19 (March 20, 2022)